Gear refactored
