+++ dicom_processor.py
@@ -2,7 +2,6 @@ import os
 import json
 import sys
 import logging
-import classification_from_label
 import flywheel
 import re
 import string
@@ -11,32 +10,23 @@ import pydicom
 import zipfile
 import pandas as pd
 import pprint
+import ast
 
-import classify_CT
-import classify_MR
-import classification_from_label
+errors_list =[]
 
-def scan_coverage_value(df):
-    df['ImagePositionPatient-Z'] = df.apply(lambda x: ast.literal_eval(x['ImagePositionPatient'])[2], axis=1)
-    max = df['ImagePositionPatient-Z'].max()
-    min = df['ImagePositionPatient-Z'].min()
-    result = max - min
-    scan_coverage = result if result > 0 else result * -1
-    return scan_coverage
+def format_string(in_string):
+    formatted = re.sub(r'[^\x00-\x7f]',r'', str(in_string)) # Remove non-ascii characters
+    formatted = ''.join(filter(lambda x: x in string.printable, formatted))
+    if len(formatted) == 1 and formatted == '?':
+        formatted = None
+    return formatted#.encode('utf-8').strip()
 
-
-# Utility:  Check a list of regexes for truthyness
-def regex_search_label(regexes, label):
-    if any(regex.search(label) for regex in regexes):
-            return True
-    else:
-            return False
 
 
 def assign_type(s):
-    """
+    '''
     Sets the type of a given input.
-    """
+    '''
     if type(s) == pydicom.valuerep.PersonName or type(s) == pydicom.valuerep.PersonName3 or type(s) == pydicom.valuerep.PersonNameBase:
         return format_string(s)
     if type(s) == list or type(s) == pydicom.multival.MultiValue:
@@ -59,13 +49,6 @@ def assign_type(s):
             except ValueError:
                 return format_string(s)
 
-
-def format_string(in_string):
-    formatted = re.sub(r'[^\x00-\x7f]',r'', str(in_string)) # Remove non-ascii characters
-    formatted = ''.join(filter(lambda x: x in string.printable, formatted))
-    if len(formatted) == 1 and formatted == '?':
-        formatted = None
-    return formatted#.encode('utf-8').strip()
 
 
 def get_seq_data(sequence, ignore_keys):
@@ -92,7 +75,47 @@ def get_seq_data(sequence, ignore_keys):
     return seq_dict
 
 
+
+def get_pydicom_header(dcm):
+    '''
+    Extract the header values
+    '''
+    header = {}
+    exclude_tags = ['[Unknown]', 'PixelData', 'Pixel Data',  '[User defined data]', '[Protocol Data Block (compressed)]', '[Histogram tables]', '[Unique image iden]']
+    tags = dcm.dir()
+    for tag in tags:
+        try:
+            if (tag not in exclude_tags) and ( type(dcm.get(tag)) != pydicom.sequence.Sequence ):
+                value = dcm.get(tag)
+                if value or value == 0: # Some values are zero
+                    # Put the value in the header
+                    if type(value) == str and len(value) < 10240: # Max pydicom field length
+                        header[tag] = format_string(value)
+                    else:
+                        header[tag] = assign_type(value)
+                else:
+                    error_message = 'No value found for tag: ' + tag
+                    errors_list.append(['debug', error_message])
+                    #log.debug('No value found for tag: ' + tag)
+
+            if type(dcm.get(tag)) == pydicom.sequence.Sequence:
+                seq_data = get_seq_data(dcm.get(tag), exclude_tags)
+                # Check that the sequence is not empty
+                if seq_data:
+                    header[tag] = seq_data
+        except:
+            error_message = 'Failed to get ' + tag
+            errors_list.append(['debug', error_message])
+            #log.debug('Failed to get ' + tag)
+            pass
+    return header
+
+
+
 def process_dicom(zip_file_path):
+    '''
+    Create Pandas Dataframe where each row is a dicom image header information
+    '''
     if zipfile.is_zipfile(zip_file_path):
         dcm_list = []
         zip_obj = zipfile.ZipFile(zip_file_path)
@@ -102,7 +125,9 @@ def process_dicom(zip_file_path):
             dcm_tmp = None
             if os.path.isfile(dcm_path):
                 try:
-                    log.info('reading %s' % dcm_path)
+                    error_message = 'reading %s' % dcm_path
+                    errors_list.append(['info', error_message])
+                    #log.info('reading %s' % dcm_path)
                     dcm_tmp = pydicom.read_file(dcm_path)
                     # Here we check for the Raw Data Storage SOP Class, if there
                     # are other pydicom files in the zip then we read the next one,
@@ -115,11 +140,15 @@ def process_dicom(zip_file_path):
                 except:
                     pass
             else:
-                log.warning('%s does not exist!' % dcm_path)
+                error_message = '%s does not exist!' % dcm_path
+                errors_list.append(['warning', error_message])
+                #log.warning('%s does not exist!' % dcm_path)
         # Select last image
         dcm = dcm_list[-1]
     else:
-        log.info('Not a zip. Attempting to read %s directly' % os.path.basename(zip_file_path))
+        error_message = 'Not a zip. Attempting to read %s directly' % os.path.basename(zip_file_path)
+        errors_list.append(['info', error_message])
+        #log.info('Not a zip. Attempting to read %s directly' % os.path.basename(zip_file_path))
         dcm = pydicom.read_file(zip_file_path)
         dcm_list = [dcm]
 
@@ -136,34 +165,5 @@ def process_dicom(zip_file_path):
         df_list.append(df_tmp)
     df = pd.concat(df_list, ignore_index=True, sort=True)
 
-    return df, dcm
-
-
-def get_pydicom_header(dcm):
-    # Extract the header values
-    header = {}
-    exclude_tags = ['[Unknown]', 'PixelData', 'Pixel Data',  '[User defined data]', '[Protocol Data Block (compressed)]', '[Histogram tables]', '[Unique image iden]']
-    tags = dcm.dir()
-    for tag in tags:
-        try:
-            if (tag not in exclude_tags) and ( type(dcm.get(tag)) != pydicom.sequence.Sequence ):
-                value = dcm.get(tag)
-                if value or value == 0: # Some values are zero
-                    # Put the value in the header
-                    if type(value) == str and len(value) < 10240: # Max pydicom field length
-                        header[tag] = format_string(value)
-                    else:
-                        header[tag] = assign_type(value)
-                else:
-                    log.debug('No value found for tag: ' + tag)
-
-            if type(dcm.get(tag)) == pydicom.sequence.Sequence:
-                seq_data = get_seq_data(dcm.get(tag), exclude_tags)
-                # Check that the sequence is not empty
-                if seq_data:
-                    header[tag] = seq_data
-        except:
-            log.debug('Failed to get ' + tag)
-            pass
-    return header
+    return df, dcm, errors_list
