+++ manifest.json
@@ -2,11 +2,11 @@
   "name": "dicom-metadata-classification",
   "label": "Metadata Classification: DICOM",
   "description": "This gear classifies DICOM images and updates the classification metadata of the input file",
-  "version": "1.0.1_sbx_dev4",
+  "version": "1.0.1_sbx_dev8",
   "custom": {
     "gear-builder": {
       "category": "converter",
-      "image": "dicom-metadata-classification:1.0.1_sbx_dev4"
+      "image": "dicom-metadata-classification:1.0.1_sbx_dev8"
     },
     "flywheel": {
       "suite": "Metadata Import and Validation"
